+++ mmaction/models/heads/__init__.py
@@ -1,3 +1,4 @@
+from .audio_tsn_head import AudioTSNHead
 from .base import BaseHead
 from .i3d_head import I3DHead
 from .slowfast_head import SlowFastHead
@@ -8,5 +9,5 @@ from .tsn_head import TSNHead
 
 __all__ = [
     'TSNHead', 'I3DHead', 'BaseHead', 'TSMHead', 'SlowFastHead', 'SSNHead',
-    'TPNHead'
+    'TPNHead', 'AudioTSNHead'
 ]