+++ mmaction/models/recognizers/__init__.py
@@ -1,5 +1,6 @@
+from .audio_recognizer import AudioRecognizer
 from .base import BaseRecognizer
 from .recognizer2d import Recognizer2D
 from .recognizer3d import Recognizer3D
 
-__all__ = ['BaseRecognizer', 'Recognizer2D', 'Recognizer3D']
+__all__ = ['BaseRecognizer', 'Recognizer2D', 'Recognizer3D', 'AudioRecognizer']